ms-user/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
mysql-connector-python==8.0.32
PyMySQL==1.0.2
Flask-SQLAlchemy==3.0.3
Flask-RESTful==0.3.9
flask-marshmallow==0.14.0